requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.6
Django==4.1.6
django-debug-toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
filelock==3.9.0
identify==2.5.18
nodeenv==1.7.0
platformdirs==3.0.0
pre-commit==3.1.1
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.19.0
